+++ lab2/src/ScaleFactor.java
@@ -3,13 +3,13 @@ import java.util.List;
 public class ScaleFactor {
 
 	
-	public float ScaleFactorCalc (float c, List<Integer> bugTime, List<Integer> bugNumber ) {
-	float k1 = 0;
-	float k2 = 0;
-	float partFirstB = 0;
-	float partOfPartFirstB = 0;
-	float partSecondB = 0;
-	float partThirdB = 0;
+	public double ScaleFactorCalc (double c, List<Integer> bugTime, List<Integer> bugNumber ) {
+		double k1 = 0;
+		double k2 = 0;
+		double partFirstB = 0;
+		double partOfPartFirstB = 0;
+		double partSecondB = 0;
+		double partThirdB = 0;
 	
 	for (int i = 0; i<bugTime.size(); i++) {
