+++ lab2/src/TestingStartTime.java
@@ -2,12 +2,12 @@ import java.util.List;
 
 public class TestingStartTime {
 	
-	public float TestingStartTimeCalc (float c, float k1, List<Integer> bugTime,  List<Integer> bugNumber) {
-	
-	float t1 = 0;
-	float partFirstG = 0;
+	public double TestingStartTimeCalc (double c, double k1, List<Integer> bugTime,  List<Integer> bugNumber) {
+
+		double t1 = 0;
+		double partFirstG = 0;
 		
-	int newc = Math.round(c);
+	int newc = (int) Math.round(c);
 		
 	for (int i = 0 ; i<(newc-bugTime.size()); i++) {
 		partFirstG = partFirstG+1/bugNumber.get(i).floatValue();
